Fixed boundary issue, it was in relation to using the wrong data type.
Changed user input to OR the data in and will clear once the information is read.
This allows for user inputs to be inputted in between game frames, current user scan rate is ~6mS.

Game logic/draw rate is supposed to be 500mS.

For my purposes, these are my game window boundaries.
YMIN 16
XMIN 8

YMAX 152
XMAX 160